services/ai-service/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.4.0
redis==5.2.1
openai==1.57.0
anthropic==0.40.0
celery[redis]==5.4.0
slowapi==0.1.9
structlog==24.4.0
httpx==0.27.2
python-dotenv==1.0.1
sse-starlette==2.1.3

# Test dependencies
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-mock==3.14.0
